🧬 [unit_test._utils.api_client] Extract the logic about getting test case data into inner modle *_test_case* for cleaning code.
