+++ test/unit_test/_utils/api_client.py
@@ -1,8 +1,5 @@
-import glob
-import os
-import pathlib
 from abc import ABCMeta, abstractmethod
-from typing import Any, List
+from typing import Any
 from unittest.mock import Mock, patch
 
 import pytest
@@ -10,23 +7,7 @@ from urllib3 import HTTPResponse, PoolManager
 
 from pymock_api._utils.api_client import BaseAPIClient, URLLibHTTPClient
 
-RESPONSE_JSON_PATHS: List[str] = []
-
-
-def _get_all_swagger_config() -> None:
-    json_dir = os.path.join(
-        str(pathlib.Path(__file__).parent.parent.parent),
-        "data",
-        "check_test",
-        "diff_with_swagger",
-        "api_response",
-        "*.json",
-    )
-    global RESPONSE_JSON_PATHS
-    for json_config_path in glob.glob(json_dir):
-        one_test_scenario = json_config_path
-        RESPONSE_JSON_PATHS.append(one_test_scenario)
-
+from ._test_case import RESPONSE_JSON_PATHS, _get_all_swagger_config
 
 _get_all_swagger_config()
